pyproject.toml
[project]
name = "ct0001-streamob"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "orjson>=3.10.16",
    "pyzmq>=26.3.0",
    "websockets>=15.0.1",
]
